Added a unittest and changed the filename to a valid format
